Update README.md
As we all are noticing that with passing time, the urge of financial freedom among the people is rising globally. So, those with minimal paycheck are inclining towards Stock Market and people with average paycheck wants to pursue the freedom through Start-up, becoming an enterpreneur, which is a good thing.
The start-ups emerging per year are more than a million, and 90% of them face failure in the early stages. 

Discussion leading to conclusion, about why a start-up fail, below reasons come into light:-
•	Weak Business Models
•	Poor Planning
•	Faulty customer insights or lack of original ideas 
•	Break in Capital chain
•	Lacking skills  or Handling single handed
•	Incompetency in finding and reaching the agreement with the investors
As for now, These reasons are the major issues for Start-up failure as well as for Entrepreneur who are steady and stable. 

So, to the above stated issues we have designed a platform as  a solution of it.





The 99 Percent
Its not only a bridge between investors and entrepreneurs but will also help them to interact with each other, for better understanding beforehead from getting involved with each other.

For Example :-
1.	Mr. X wants to be a entrepreneur with the idea he’s having for A start-up but he lacks in skill.
2.	Mr. Y also wants to be a entrepreneur, he’s very skilled but not having any unique idea.
3.	Mr. E is a successful entrepreneur, but having some problem with capital chain for his company.
4.	Mr. N is a young investor, he lacks experience in selection but wants to start investing in start-ups.
5.	Mr. T is an experienced investor, he’s experienced and has invested in many successful start-ups.

So, here on this platform Mr. X and Mr. Y can interact with each other to form a team(NAME:- R), and reach Mr. N for investment, as Mr. N has no experience in this field so he can seek advice from Mr. T, And both the investors can decide what to do(whether to invest or not).
Team R is new in this field, so they canask Mr. E, to be their mentor.
As Mr. E is having some issues concerning the Capital Supply  so he can contact Mr. T for Investing in his company.

So in this way The business world can build a strong in it.
It will also decrease the domination of the venture capital firm.
It will also help in increasing the success rate of start-ups.